requirements.txt
langchain-google-genai==0.0.5
python-dotenv==1.0.0
flask==3.0.0